Splitting API calls in BrowsePetsView into Controller
