+++ src/app/BrowsePets/BrowsePetsController.js
@@ -1,10 +1,12 @@
 "use client"
 
 import React, {useState, useEffect} from "react";
-import {getAllAvailablePets} from "@/app/BrowsePets/BrowsePetsAPI";
+import {
+  getAllAvailablePets,
+  getAllAvailablePetsOfType,
+} from "@/app/BrowsePets/BrowsePetsAPI";
 import Animal from "@/app/BrowsePets/Animal";
 import BrowsePetsView from "@/app/BrowsePets/BrowsePetsView";
-
 
 const BrowsePetsController = () => {
   const [animals, SetAnimals] = useState([]);
@@ -26,6 +28,32 @@ const BrowsePetsController = () => {
   }, []);
 
   return <BrowsePetsView animals={animals} />;
+};
+
+export const fetchAnimalsByCategory = async (currentCategory, animal) => {
+  try {
+    let animals;
+    if (currentCategory === animal) {
+      animals = await getAllAvailablePets();
+    } else if (currentCategory) {
+      animals = await getAllAvailablePetsOfType(currentCategory);
+    }
+    // create animal objects for each animal and store in state
+    const animalObjects = animals.map((animal) => {
+      return new Animal(
+        animal.pet_id,
+        animal.name,
+        animal.age,
+        animal.sex,
+        animal.category,
+        animal.breed,
+        ""
+      );
+    });
+    return animalObjects;
+  } catch (error) {
+    console.error("Error fetching animals by category:", error);
+  }
 };
 
 export default BrowsePetsController;
